assets/requirements.txt
pyserial
opencv-python-headless
pillow
pyzbar
numpy
scipy
pandas
openpyxl